% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecc.R
\name{apply_rank}
\alias{apply_rank}
\title{Rank ensemble members.}
\usage{
apply_rank(x, origin, ...)
}
\arguments{
\item{x}{An ensemble to be ranked.}

\item{origin}{An ensemble providing the ranks to be applied.}

\item{...}{Not used yet.}
}
\description{
Take the rank structure from the \code{origin} ensemble and
apply it to the \code{x} ensemble.
}
