% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_vec_recyclable}
\alias{are_vec_recyclable}
\title{Are vectors recyclable?}
\usage{
are_vec_recyclable(...)
}
\arguments{
\item{...}{Vectors to be tested for recycling.}
}
\value{
A scalar logical: \code{TRUE} or \code{FALSE}.
}
\description{
Are vectors recyclable?
}
\keyword{internal}
