% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_class.R
\name{as_ensemble}
\alias{as_ensemble}
\alias{as_ensemble.default}
\title{Coerce object to ensemble.}
\usage{
as_ensemble(x, ...)

\method{as_ensemble}{default}(x, ...)
}
\arguments{
\item{x}{An object.}
}
\value{
An object of class ensemble.
}
\description{
Generic function to coerce objects to objects of class ensemble.
}
\details{
The default methods applies \code{as.matrix}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method.
}}

