% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{code_to_strand}
\alias{code_to_strand}
\title{Convert strand integer codes to strand words}
\usage{
code_to_strand(code, .default = NA_character_, .missing = NA_character_)
}
\arguments{
\item{code}{Integer vector of strandness codes: \code{1L} or
\code{-1L}.}

\item{.default}{What value to use by default for values other than
\code{1L} or \code{-1L}.}

\item{.missing}{What value to use when recoding \code{NA_integer_}.}
}
\value{
A character vector.
}
\description{
This function converts an integer vector of values \code{1L} or \code{-1L} to
\code{'forward'} or \code{'reverse'}, respectively.
}
\seealso{
\code{\link{strand_to_code}}
}
\keyword{internal}
