% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{._get_divisions}
\alias{._get_divisions}
\title{Get list of all Ensembl divisions}
\usage{
._get_divisions(callback = "randomlygeneratedname")
}
\arguments{
\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}
}
\value{
A parsed JSON response containing the list of Ensembl divisions.

See more about the implemented endpoint \code{\link[=._get_divisions]{._get_divisions()}}
on the following \href{https://rest.ensembl.org/documentation/info/info_divisions}{GET info/divisions}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of all available Ensembl divisions for which information is accessible.
}
\examples{
._get_divisions()
._get_divisions(callback = "randomlygeneratedname")
}
