% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{._get_rest_version}
\alias{._get_rest_version}
\title{Get REST API Version}
\usage{
._get_rest_version(callback = "randomlygeneratedname")
}
\arguments{
\item{callback}{(Optional) A string representing the name of the callback
subroutine for JSONP responses.}
}
\value{
A parsed JSON response containing the REST API version information.

See more about the implemented endpoint \code{\link[=._get_rest_version]{._get_rest_version()}}
on the following \href{https://rest.ensembl.org/documentation/info/rest}{GET info/rest}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves the current version of the Ensembl REST API.
}
\examples{
._get_rest_version()
._get_rest_version(callback = "randomlygeneratedname")
}
