% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{._get_software_version}
\alias{._get_software_version}
\title{Get Software Version}
\usage{
._get_software_version(callback = "randomlygeneratedname")
}
\arguments{
\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}
}
\value{
A parsed JSON response containing the Ensembl API version information.

See more about the implemented endpoint \code{\link[=._get_software_version]{._get_software_version()}}
on the following \href{https://rest.ensembl.org/documentation/info/software}{GET info/software}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves the current version of the Ensembl API used by the REST server.
}
\examples{
._get_software_version()
._get_software_version(callback = "randomlygeneratedname")
}
