% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{._get_variation_sources}
\alias{._get_variation_sources}
\title{Get Variation Sources for a Species}
\usage{
._get_variation_sources(
  species,
  callback = "randomlygeneratedname",
  filter = NULL
)
}
\arguments{
\item{species}{(Required) A string representing the species name or alias (e.g., "homo_sapiens").}

\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}

\item{filter}{(Optional) A string to restrict the variation source searches to a single source
(e.g., "dbSNP", "ClinVar", "OMIM", "UniProt", "HGMD").}
}
\value{
A parsed JSON response containing the variation sources for the specified species.

See more about the implemented endpoint \code{\link[=._get_variation_sources]{._get_variation_sources()}}
on the following \href{https://rest.ensembl.org/documentation/info/variation}{GET info/variation/:species}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves the variation sources used in Ensembl for a given species.
}
\examples{
._get_variation_sources("homo_sapiens")
._get_variation_sources("homo_sapiens", filter = "ClinVar")
._get_variation_sources("homo_sapiens", callback = "randomlygeneratedname")
}
