% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecc.R
\name{ecc_q}
\alias{ecc_q}
\alias{ecc_r}
\title{Ensemble copula coupling.}
\usage{
ecc_q(model, newdata, origin_name, ...)

ecc_r(model, newdata, origin_name, ...)
}
\arguments{
\item{model}{A fitted model object.}

\item{newdata}{A data.frame with new data to predict on.}

\item{origin_name}{A character. The name of the \code{ensemble} variable in
\code{newdata}, from which the ranks will be applied to the predicted
quantiles.}

\item{...}{Not used yet.}
}
\description{
Apply empirical ensemble copula coupling.
The functions wrap the predicting of quantiles and re-ordering of ranks.
}
