% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_transformer.R
\name{ens_sd}
\alias{ens_sd}
\title{Ensemble standard deviation}
\usage{
ens_sd(x, ...)
}
\arguments{
\item{x}{An ensemble.}

\item{...}{passed to \code{sd}.}
}
\description{
Compute the ensemble standard deviation.
}
\details{
Apply \code{sd} on each row of the ensemble.
}
