% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{ensembl_server}
\alias{ensembl_server}
\title{Ensembl REST API server}
\usage{
ensembl_server()
}
\value{
A string containing Ensembl REST API server URL.
}
\description{
Ensembl REST API server
}
\keyword{internal}
