% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblr-package.R
\docType{package}
\name{ensemblr-package}
\alias{ensemblr}
\alias{ensemblr-package}
\title{ensemblr: R Client for the Ensembl REST API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R Client for the Ensembl REST API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramiromagno/ensemblr}
  \item \url{http://rmagno.eu/ensemblr/}
  \item Report bugs at \url{https://github.com/ramiromagno/ensemblr/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Isabel Duarte \email{iduarte.scientist@gmail.com} (\href{https://orcid.org/0000-0003-0060-2936}{ORCID})
  \item Ana-Teresa Maia \email{maia.anateresa@gmail.com} (\href{https://orcid.org/0000-0002-0454-9207}{ORCID})
}

Other contributors:
\itemize{
  \item CINTESIS [copyright holder, funder]
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
