% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_range.R
\name{genomic_range}
\alias{genomic_range}
\title{Create genomic range strings}
\usage{
genomic_range(chr, start, end, starting_position_index = 1L)
}
\arguments{
\item{chr}{A character vector of chromosome names.}

\item{start}{An integer vector of start positions.}

\item{end}{An integer vector of end positions.}

\item{starting_position_index}{Use this argument to indicate if the positions
are 0-based (\code{0L}) or 1-based (\code{1L}). This value is used to check
if positions are equal or above this number.}
}
\value{
Returns a character vector whose strings are genomic ranges of the
form \{chr\}:\{start\}..\{end\}.
}
\description{
This function converts three vectors: \code{chr}, \code{start}, and \code{end}
to strings of the form \{chr\}:\{start\}..\{end\}.
}
\examples{
genomic_range("1", 10000L, 20000L) # Returns "1:10000..20000"

}
