% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_alignment_by_region}
\alias{get_alignment_by_region}
\title{Get alignment by region}
\usage{
get_alignment_by_region(species, region)
}
\arguments{
\item{species}{A string representing the species name (e.g., "homo_sapiens").}

\item{region}{A string representing the genomic region (e.g., "3:1000-2000").}
}
\value{
A list of parsed JSON responses containing the genomic alignments
for the provided species and region.

See more about the implemented endpoint \code{\link[=get_alignment_by_region]{get_alignment_by_region()}}
on the following \href{https://rest.ensembl.org/documentation/info/genomic_alignment_region}{GET alignment/region/:species/:region}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves genomic alignments as separate blocks based on a region and species
}
\examples{
get_alignment_by_region("homo_sapiens", "3:1000-2000")
}
