% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_analyses}
\alias{get_analyses}
\title{Get analyses behind Ensembl databases}
\usage{
get_analyses(species_name, verbose = FALSE, warnings = TRUE,
  progress_bar = TRUE)
}
\arguments{
\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Examples: \code{'homo_sapiens'}
(human), \code{'ovis_aries'} (Domestic sheep) or \code{'capra_hircus'} (Goat).}

\item{verbose}{Whether to be verbose about the http requests and respective
responses' status.}

\item{warnings}{Whether to show warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}} of 3 variables:
\describe{
  \item{\code{species_name}}{Ensembl species name: this is the name used
  internally by Ensembl to uniquely identify a species by name. It is the
  scientific name but formatted without capitalisation and spacing converted
  with an underscore, e.g., \code{'homo_sapiens'}.}
  \item{\code{database}}{Ensembl database. Typically one of \code{'core'},
  \code{'rnaseq'}, \code{'cdna'}, \code{'funcgen'} and
  \code{'otherfeatures'}.}
  \item{\code{analysis}}{Analysis.}
}
}
\description{
This function retrieves a table of analyses involved in the generation of
data for the different Ensembl databases.
}
