% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_analysis_info}
\alias{get_analysis_info}
\title{Get the names of analyses involved in generating Ensembl data}
\usage{
get_analysis_info(species, callback = "randomlygeneratedname")
}
\arguments{
\item{species}{A string representing the species name or alias
(e.g., "homo_sapiens").}

\item{callback}{(Optional) A string representing the name of the callback
subroutine for JSONP responses.}
}
\value{
A list of analysis names related to the specified species.

See more about the implemented endpoint \code{\link[=get_analysis_info]{get_analysis_info()}}
on the following \href{https://rest.ensembl.org/documentation/info/analysis}{GET info/analysis/:species}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of analysis names associated with generating Ensembl
data for a given species.
}
\examples{
get_analysis_info("homo_sapiens")
get_analysis_info("homo_sapiens", callback = "randomlygeneratedname")
}
