% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assemblies.R
\name{get_assemblies}
\alias{get_assemblies}
\title{Get details about the genome assembly of a species}
\usage{
get_assemblies(
  species_name = "homo_sapiens",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Examples: \code{'homo_sapiens'}
(human), \code{'ovis_aries'} (Domestic sheep) or \code{'capra_hircus'}
(Goat).}

\item{verbose}{Whether to be chatty.}

\item{warnings}{Whether to print warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}}, each row being a toplevel sequence,
of 4 variables:
\describe{
\item{\code{species_name}}{Ensembl species name: this is the name used internally
by Ensembl to uniquely identify a species by name. It is the scientific
name but formatted without capitalisation and spacing converted with an
underscore, e.g., \code{'homo_sapiens'}.}
\item{\code{assembly_name}}{Assembly name.}
\item{\code{assembly_date}}{Assembly date.}
\item{\code{genebuild_method}}{Annotation method.}
\item{\code{golden_path_length}}{Golden path length.}
\item{\code{genebuild_initial_release_date}}{Genebuild release date.}
\item{\code{default_coord_system_version}}{Default coordinate system version.}
\item{\code{assembly_accession}}{Assembly accession.}
\item{\code{genebuild_start_date}}{Genebuild start date.}
\item{\code{genebuild_last_geneset_update}}{Genebuild last geneset update.}
}
}
\description{
This functions retrieves details about the assembly of a queried species.
}
\examples{
# Get details about the human assembly
get_assemblies()

# Get details about the Mouse and the Fruit Fly genomes
get_assemblies(c('mus_musculus', 'drosophila_melanogaster'))

}
