% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_assembly_info}
\alias{get_assembly_info}
\title{Get the available assemblies for a species}
\usage{
get_assembly_info(
  species,
  bands = 0,
  synonyms = 0,
  callback = "randomlygeneratedname"
)
}
\arguments{
\item{species}{A string representing the species name or alias
(e.g., "homo_sapiens").}

\item{bands}{(Optional) A boolean (0 or 1) indicating whether to
include karyotype band information. Default is 0.}

\item{synonyms}{(Optional) A boolean (0 or 1) indicating whether to
include information about known synonyms. Default is 0.}

\item{callback}{(Optional) A string representing the name of the callback
subroutine for JSONP responses.}
}
\value{
A list of parsed JSON responses containing information about
the available assemblies for the specified species.

See more about the implemented endpoint \code{\link[=get_assembly_info]{get_assembly_info()}}
on the following \href{https://rest.ensembl.org/documentation/info/assembly_info}{GET info/assembly/:species}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of available assemblies for a given species,
including toplevel sequences, chromosomes, and optionally cytogenetic
bands and synonyms.
}
\examples{
get_assembly_info("homo_sapiens")
get_assembly_info("homo_sapiens", bands = 1)
get_assembly_info("homo_sapiens", synonyms = 1,
  callback = "randomlygeneratedname")
}
