% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_biotypes}
\alias{get_biotypes}
\title{Get the functional classifications of gene models for a species}
\usage{
get_biotypes(species, callback = "randomlygeneratedname")
}
\arguments{
\item{species}{A string representing the species name or alias (e.g., "homo_sapiens").}

\item{callback}{(Optional) A string representing the name of the callback subroutine
for JSONP responses.}
}
\value{
A list of parsed JSON responses containing the biotypes for the specified species.

See more about the implemented endpoint \code{\link[=get_biotypes]{get_biotypes()}}
on the following \href{https://rest.ensembl.org/documentation/info/biotypes}{GET info/biotypes/:species}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves the list of functional classifications (biotypes) of gene models
that Ensembl associates with a particular species.
Useful for restricting the type of genes/transcripts retrieved by other endpoints.
}
\examples{
get_biotypes("homo_sapiens")
get_biotypes("homo_sapiens", callback = "randomlygeneratedname")
}
