% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_biotypes_by_name}
\alias{get_biotypes_by_name}
\title{Get properties of biotypes by name}
\usage{
get_biotypes_by_name(
  name,
  object_type = "",
  callback = "randomlygeneratedname"
)
}
\arguments{
\item{name}{A string representing the biotype name (e.g., "protein_coding").}

\item{object_type}{(Optional) A string specifying the object type ("gene" or "transcript").}

\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}
}
\value{
A list of parsed JSON responses containing the properties of biotypes with the given name.

See more about the implemented endpoint \code{\link[=get_biotypes_by_name]{get_biotypes_by_name()}}
on the following \href{https://rest.ensembl.org/documentation/info/biotypes_name}{GET info/biotypes/name/:name/:object_type}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves the properties of biotypes with a given name. Optionally, the object type (gene or transcript) can be provided for filtering.
}
\examples{
get_biotypes_by_name("protein_coding")
get_biotypes_by_name("protein_coding", object_type = "gene")
get_biotypes_by_name("protein_coding", object_type = "gene", callback = "randomlygeneratedname")
}
