% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_biotypes_groups}
\alias{get_biotypes_groups}
\title{Get properties of biotypes within a group}
\usage{
get_biotypes_groups(
  group = "",
  object_type = "",
  callback = "randomlygeneratedname"
)
}
\arguments{
\item{group}{(Optional) A string representing the biotype group (e.g., "coding").
If not provided, the available biotype groups are returned.}

\item{object_type}{(Optional) A string specifying the object type ("gene" or "transcript").}

\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}
}
\value{
A list of parsed JSON responses containing the properties of biotypes within
the specified group or all biotype groups if no group is provided.

See more about the implemented endpoint \code{\link[=get_biotypes_groups]{get_biotypes_groups()}}
on the following \href{https://rest.ensembl.org/documentation/info/biotypes_groups}{GET info/biotypes/groups/:group/:object_type}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of available biotype groups or, if a group is specified,
the properties of biotypes within that group.
Optionally, the object type (gene or transcript) can be used to filter the results.
}
\examples{
get_biotypes_groups()
get_biotypes_groups(group = "coding")
get_biotypes_groups(group = "coding", object_type = "gene")
get_biotypes_groups(group = "coding", object_type = "gene", callback = "randomlygeneratedname")
}
