% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_cafe_genetree_by_species_id}
\alias{get_cafe_genetree_by_species_id}
\title{Get cafe gene tree by species id}
\usage{
get_cafe_genetree_by_species_id(species, id)
}
\arguments{
\item{species}{A string representing the species name (e.g., "homo_sapiens").}

\item{id}{A string representing the gene, transcript, or translation
stable identifier.}
}
\value{
A list of parsed JSON responses containing the cafe tree for
the provided species and stable identifier.
}
\description{
Retrieves the cafe tree of the gene tree that contains the
gene/transcript/translation stable identifier in the given species
}
\note{
See more about the implemented endpoint \code{\link[=get_cafe_genetree_by_species_id]{get_cafe_genetree_by_species_id()}}
on the following \href{https://rest.ensembl.org/documentation/info/cafe_tree_species_member_id}{GET cafe/genetree/member/id/:species/:id}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\examples{
get_cafe_genetree_by_species_id("homo_sapiens", "ENST00000380152")
}
