% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_compara_methods}
\alias{get_compara_methods}
\title{Get list of compara methods}
\usage{
get_compara_methods(
  callback = "randomlygeneratedname",
  class = NULL,
  compara = NULL
)
}
\arguments{
\item{callback}{(Optional) A string representing the name of the callback subroutine
for JSONP responses.}

\item{class}{(Optional) A string specifying the class of the method to query for.
Regular expression patterns are supported (e.g., "GenomicAlign").}

\item{compara}{(Optional) A string representing the name of the compara database
to use (e.g., "vertebrates").}
}
\value{
A list of parsed JSON responses containing all available compara methods.

See more about the implemented endpoint \code{\link[=get_compara_methods]{get_compara_methods()}}
on the following \href{https://rest.ensembl.org/documentation/info/compara_methods}{GET info/compara/methods}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of all compara analyses available (an analysis defines the type
of comparative data). Optional filtering by class or compara database can be applied.
}
\examples{
get_compara_methods()
get_compara_methods(class = "GenomicAlign")
get_compara_methods(compara = "vertebrates", class = "GenomicAlign")
}
