% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_compara_species_sets}
\alias{get_compara_species_sets}
\title{Get collections of species analysed with a specified compara method}
\usage{
get_compara_species_sets(
  method,
  callback = "randomlygeneratedname",
  compara = NULL
)
}
\arguments{
\item{method}{A string representing the compara method to filter by (e.g., "EPO").}

\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}

\item{compara}{(Optional) A string representing the name of the compara database to use (e.g., "vertebrates").}
}
\value{
A list of parsed JSON responses containing all collections of species analysed with the specified compara method.

See more about the implemented endpoint \code{\link[=get_compara_species_sets]{get_compara_species_sets()}}
on the following \href{https://rest.ensembl.org/documentation/info/compara_species_sets}{GET info/compara/species_sets/:method}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of all collections of species analysed with the specified compara method.
The compara method must be one of the methods returned by the \verb{/info/compara/methods} endpoint.
}
\examples{
get_compara_species_sets("EPO")
get_compara_species_sets("EPO", compara = "vertebrates")
}
