% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_comparas}
\alias{get_comparas}
\title{Get a list of all available comparative genomics databases and their data release}
\usage{
get_comparas(callback = "randomlygeneratedname")
}
\arguments{
\item{callback}{(Optional) A string representing the name of the callback subroutine
for JSONP responses.}
}
\value{
A list of parsed JSON responses containing all available comparative genomics
databases and their data release.

See more about the implemented endpoint \code{\link[=get_comparas]{get_comparas()}}
on the following \href{https://rest.ensembl.org/documentation/info/comparas}{GET info/comparas}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of all available comparative genomics databases and their data release.
This endpoint is deprecated, and users are advised to use the \verb{/info/genomes/division} endpoint instead.
}
\examples{
get_comparas()
}
