% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_consequence_types}
\alias{get_consequence_types}
\title{Get Variant Consequence Types}
\usage{
get_consequence_types(callback = "randomlygeneratedname", rank = NULL)
}
\arguments{
\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}

\item{rank}{(Optional) A boolean (0 or 1) to include consequence ranking in the response. Default is 0.}
}
\value{
A parsed JSON response containing the list of variant consequence types.

See more about the implemented endpoint \code{\link[=get_consequence_types]{get_consequence_types()}}
on the following \href{https://rest.ensembl.org/documentation/info/variation_consequence_types}{GET info/variation/consequence_types}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of all variant consequence types available in Ensembl.
}
\examples{
get_consequence_types()
get_consequence_types(rank = 1)
get_consequence_types(callback = "randomCallback")
}
