% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cytogenetic_bands.R
\name{get_cytogenetic_bands}
\alias{get_cytogenetic_bands}
\title{Get cytogenetic bands by species}
\usage{
get_cytogenetic_bands(
  species_name = "homo_sapiens",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Examples: \code{'homo_sapiens'}
(human), \code{'ovis_aries'} (Domestic sheep) or \code{'capra_hircus'}
(Goat).}

\item{verbose}{Whether to be chatty.}

\item{warnings}{Whether to print warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}}, each row being a cytogenetic band,
of 8 variables:
\describe{
\item{\code{species_name}}{Ensembl species name: this is the name used internally
by Ensembl to uniquely identify a species by name. It is the scientific
name but formatted without capitalisation and spacing converted with an
underscore, e.g., \code{'homo_sapiens'}.}
\item{\code{assembly_name}}{Assembly name.}
\item{\code{cytogenetic_band}}{Name of the cytogenetic_band.}
\item{\code{chromosome}}{Chromosome name.}
\item{\code{start}}{Genomic start position of the cytogenetic band. Starts at 1.}
\item{\code{end}}{Genomic end position of the cytogenetic band. End position is
included in the band interval.}
\item{\code{stain}}{\href{https://en.wikipedia.org/wiki/Giemsa_stain}{Giemsa
stain} results: Giemsa negative, \code{'gneg'}; Giemsa positive, of
increasing intensities, \code{'gpos25'}, \code{'gpos50'}, \code{'gpos75'},
and \code{'gpos100'}; centromeric region, \code{'acen'}; heterochromatin,
either pericentric or telomeric, \code{'gvar'}; and short arm of
acrocentric chromosomes are coded as \code{'stalk'}.}
\item{\code{strand}}{Strand.}
}
}
\description{
This function retrieves cytogenetic bands. If no cytogenetic information is
available for the queried species then it will be omitted from in the
returned value.
}
\examples{
# Get toplevel sequences for the human genome (default)
get_cytogenetic_bands()

# Get toplevel sequences for Mus musculus
get_cytogenetic_bands('mus_musculus')

}
