% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_data}
\alias{get_data}
\title{Get a list of available data releases on the Ensembl REST server}
\usage{
get_data(callback = "randomlygeneratedname")
}
\arguments{
\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}
}
\value{
A list of parsed JSON responses containing the available data releases on the Ensembl REST server.

See more about the implemented endpoint \code{\link[=get_data]{get_data()}}
on the following \href{https://rest.ensembl.org/documentation/info/data}{GET info/data}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of the data releases available on the Ensembl REST server. It may return more than one release if the server has an unfrequent, non-standard configuration.
}
\examples{
get_data()
}
