% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versioning.R
\name{get_data_versions}
\alias{get_data_versions}
\title{Retrieve the data release version(s) available on the Ensembl REST server.}
\usage{
get_data_versions(verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{verbose}{Whether to be chatty.}

\item{warnings}{Whether to print warnings.}
}
\value{
An integer vector of release version(s).
}
\description{
Retrieve the data release version(s) available on the Ensembl REST server.
}
