% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_eg_version}
\alias{get_eg_version}
\title{Get the Ensembl Genomes version of the databases backing the service}
\usage{
get_eg_version(callback = "randomlygeneratedname")
}
\arguments{
\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}
}
\value{
A parsed JSON response containing the Ensembl Genomes version of the databases.

See more about the implemented endpoint \code{\link[=get_eg_version]{get_eg_version()}}
on the following \href{https://rest.ensembl.org/documentation/info/eg_version}{GET info/eg_version}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves the Ensembl Genomes version of the databases supporting the current service.
}
\examples{
get_eg_version()
}
