% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ensembl_genomes_version.R
\name{get_ensembl_genomes_version}
\alias{get_ensembl_genomes_version}
\title{Get Ensembl Genomes version}
\usage{
get_ensembl_genomes_version(verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{verbose}{Whether to be chatty.}

\item{warnings}{Whether to print warnings.}
}
\value{
An integer value: the Ensembl Genomes version.
}
\description{
Returns the Ensembl Genomes version of the databases backing this service.
}
\examples{
get_ensembl_genomes_version()

}
