% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eqtl_pval_by_gene.R
\name{get_eqtl_pval_by_gene}
\alias{get_eqtl_pval_by_gene}
\title{Get eQTL details by gene ensembl identifier}
\usage{
get_eqtl_pval_by_gene(
  ensembl_id,
  species_name = "homo_sapiens",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{ensembl_id}{An Ensembl stable identifier, e.g. \code{"ENSG00000248234378"}
(or a vector thereof).}

\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Currently, only
\code{'homo_sapiens'} (human) is supported.}

\item{verbose}{Whether to be verbose about the http requests and respective
responses' status.}

\item{warnings}{Whether to show warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}} of 10
variables:
\describe{
\item{\code{species_name}}{Ensembl species name: this is the name used
internally by Ensembl to uniquely identify a species by name. It is the
scientific name but formatted without capitalisation and spacing converted
with an underscore, e.g., \code{'homo_sapiens'}.}
\item{\code{ensembl_id}}{An Ensembl stable identifier, e.g.
\code{"ENSG00000248234378"}.}
\item{\code{variant_id}}{Variant identifier, e.g. \code{"rs80100814"}.}
\item{\code{tissue}}{Tissue.}
\item{\code{display_consequence}}{Variant consequence type.}
\item{\code{seq_region_name}}{Sequence region name (typically a chromosome name)
of the variant.}
\item{\code{seq_region_start}}{Genomic start position of the variant.}
\item{\code{seq_region_end}}{Genomic end position of the variant.}
\item{\code{beta}}{The effect size beta of the eQTL association analysis.}
\item{\code{pvalue}}{The p-value of the eQTL association analysis.}
}
}
\description{
This function retrieves eQTLs, along with its genomic position, beta values
and p-values.
}
\section{Ensembl REST API endpoints}{
\code{get_eqtl_pval_by_gene()} makes GET requests to
\href{https://rest.ensembl.org/documentation/info/species_id}{/eqtl/id/:species/:stable_id}.
}

\examples{
get_eqtl_pval_by_gene('ENSG00000248378')

}
