% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eqtl_pval_by_variant.R
\name{get_eqtl_pval_by_variant}
\alias{get_eqtl_pval_by_variant}
\title{Get eQTL details by variant identifier}
\usage{
get_eqtl_pval_by_variant(
  variant_id,
  species_name = "homo_sapiens",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{variant_id}{A variant identifier, e.g. \code{"rs80100814"}.}

\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Currently, only
\code{'homo_sapiens'} (human) is supported.}

\item{verbose}{Whether to be verbose about the http requests and respective
responses' status.}

\item{warnings}{Whether to show warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}} of 6
variables:
\describe{
\item{\code{species_name}}{Ensembl species name: this is the name used
internally by Ensembl to uniquely identify a species by name. It is the
scientific name but formatted without capitalisation and spacing converted
with an underscore, e.g., \code{'homo_sapiens'}.}
\item{\code{variant_id}}{Variant identifier, e.g. \code{"rs80100814"}.}
\item{\code{ensembl_id}}{An Ensembl stable identifier, e.g.
\code{"ENSG00000248234378"}.}
\item{\code{tissue}}{Tissue.}
\item{\code{beta}}{The effect size beta of the eQTL association analysis.}
\item{\code{pvalue}}{The p-value of the eQTL association analysis.}
}
}
\description{
This function retrieves genes (Ensembl identifier) along with beta values and
p-values for the eQTL association.
}
\section{Ensembl REST API endpoints}{
\code{get_eqtl_pval_by_variant()} makes GET requests to
\href{https://rest.ensembl.org/documentation/info/species_variant}{/eqtl/variant_name/:species/:variant_name}.
}

\examples{
get_eqtl_pval_by_variant('rs80100814')

}
