% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl.R
\name{get_eqtl_tissues}
\alias{get_eqtl_tissues}
\title{Get tissues in the eQTL database}
\usage{
get_eqtl_tissues(
  species_name = "homo_sapiens",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Currently, only human
\code{'homo_sapiens'} is available.}

\item{verbose}{Whether to be verbose about the http requests and respective
responses' status.}

\item{warnings}{Whether to show warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\description{
This function retrieves all tissues in the eQTL database.
}
\section{Ensembl REST API endpoints}{
\code{get_eqtl_tissues()} makes GET requests to
\href{https://rest.ensembl.org/documentation/info/tissues}{/eqtl/tissue/:species/}.
}

\examples{
get_eqtl_tissues()

}
