% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_external_dbs}
\alias{get_external_dbs}
\title{Get external databases for a species}
\usage{
get_external_dbs(
  species,
  callback = "randomlygeneratedname",
  feature = NULL,
  filter = NULL
)
}
\arguments{
\item{species}{A string representing the species name or alias (e.g., "homo_sapiens").}

\item{callback}{(Optional) A string representing the name of the callback subroutine
for JSONP responses.}

\item{feature}{(Optional) A string representing the feature to filter external DB entries
(e.g., "dna_align_feature", "protein_align_feature", "unmapped_object", "xref", "seq_region_synonym").}

\item{filter}{(Optional) A string to restrict external DB searches to a single source
or pattern (e.g., "HGNC", "GO\%").}
}
\value{
A parsed JSON response containing a list of external sources associated
with the given species.

See more about the implemented endpoint \code{\link[=get_external_dbs]{get_external_dbs()}}
on the following \href{https://rest.ensembl.org/documentation/info/external_dbs}{GET info/external_dbs/:species}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of all available external sources for a given species.
}
\examples{
get_external_dbs("homo_sapiens")
get_external_dbs("homo_sapiens", feature = "xref", filter = "HGNC")
}
