% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_genetree_by_id}
\alias{get_genetree_by_id}
\title{Get gene tree by id}
\usage{
get_genetree_by_id(id)
}
\arguments{
\item{id}{A string representing the gene tree stable identifier.}
}
\value{
A list of parsed JSON responses containing the gene tree for
the provided gene tree stable identifier.
}
\description{
Retrieves a gene tree for a gene tree stable identifier
}
\note{
See more about the implemented endpoint \code{\link[=get_genetree_by_id]{get_genetree_by_id()}}
on the following \href{https://rest.ensembl.org/documentation/info/genetree}{GET genetree/id/:id}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\examples{
get_genetree_by_id("ENSGT00390000003602")
}
