% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_genetree_by_species_id}
\alias{get_genetree_by_species_id}
\title{Get gene tree by species id}
\usage{
get_genetree_by_species_id(species, id)
}
\arguments{
\item{species}{A string representing the species name (e.g., "homo_sapiens").}

\item{id}{A string representing the gene, transcript, or translation stable
identifier.}
}
\value{
A list of parsed JSON responses containing the gene tree
for the provided species and stable identifier.
}
\description{
Retrieves the gene tree that contains the gene/transcript/translation
stable identifier in the given species
}
\note{
See more about the implemented endpoint \code{\link[=get_genetree_by_species_id]{get_genetree_by_species_id()}}
on the following \href{https://rest.ensembl.org/documentation/info/genetree_species_member_id}{GET genetree/member/id/:species/:id}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\examples{
get_genetree_by_species_id("homo_sapiens", "ENST00000380152")
}
