% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_genome_info}
\alias{get_genome_info}
\title{Get genome information}
\usage{
get_genome_info(name, callback = "randomlygeneratedname", expand = NULL)
}
\arguments{
\item{name}{(Required) A string representing the production name of the genome
(e.g., "arabidopsis_thaliana").}

\item{callback}{(Optional) A string representing the name of the callback subroutine
for JSONP responses.}

\item{expand}{(Optional) A boolean value (0 or 1). If set to 1, expands
the information to include details of sequences (can be very large).
Default is NULL.}
}
\value{
A parsed JSON response containing information about the specified genome.

See more about the implemented endpoint \code{\link[=get_genome_info]{get_genome_info()}}
on the following \href{https://rest.ensembl.org/documentation/info/info_genome}{GET info/genomes/:genome_name}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves detailed information about a given genome based on its production name.
}
\examples{
get_genome_info(name = "arabidopsis_thaliana")
get_genome_info(name = "arabidopsis_thaliana", expand = 1)
get_genome_info(name = "arabidopsis_thaliana", callback = "randomlygeneratedname")
}
