% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_genome_info_by_accession}
\alias{get_genome_info_by_accession}
\title{Get genome information by INSDC accession}
\usage{
get_genome_info_by_accession(
  accession,
  callback = "randomlygeneratedname",
  expand = NULL
)
}
\arguments{
\item{accession}{(Required) A string representing the INSDC sequence accession (optionally versioned), e.g., "U00096".}

\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}

\item{expand}{(Optional) A boolean value (0 or 1). If set to 1, expands the information to include details of sequences (can be very large).}
}
\value{
A parsed JSON response containing information about genomes with the specified INSDC accession.

See more about the implemented endpoint \code{\link[=get_genome_info_by_accession]{get_genome_info_by_accession()}}
on the following \href{https://rest.ensembl.org/documentation/info/info_genomes_accession}{GET info/genomes/accession/:accession}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves detailed information about genomes containing a specified INSDC accession.
}
\examples{
get_genome_info_by_accession(accession = "U00096")
get_genome_info_by_accession(accession = "U00096", expand = 1)
get_genome_info_by_accession(accession = "U00096", callback = "randomlygeneratedname")
}
