% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_genome_info_by_taxonomy}
\alias{get_genome_info_by_taxonomy}
\title{Get genome information by taxonomy node}
\usage{
get_genome_info_by_taxonomy(
  taxon_name,
  callback = "randomlygeneratedname",
  expand = NULL
)
}
\arguments{
\item{taxon_name}{(Required) A string representing the taxon name or NCBI taxonomy ID (e.g., "Homo sapiens").}

\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}

\item{expand}{(Optional) A boolean value (0 or 1). If set to 1, expands the information
to include details of sequences (can be very large).}
}
\value{
A parsed JSON response containing information about genomes beneath the specified taxonomy node.

See more about the implemented endpoint \code{\link[=get_genome_info_by_taxonomy]{get_genome_info_by_taxonomy()}}
on the following \href{https://rest.ensembl.org/documentation/info/info_genomes_taxonomy}{GET info/genomes/taxonomy/:taxon_name}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves information about all genomes beneath a given node of the taxonomy.
}
\examples{
get_genome_info_by_taxonomy(taxon_name = "Homo sapiens")
get_genome_info_by_taxonomy(taxon_name = "Homo sapiens", expand = 1)
get_genome_info_by_taxonomy(taxon_name = "Homo sapiens", callback = "randomlygeneratedname")
}
