% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_homology_by_species_id}
\alias{get_homology_by_species_id}
\title{Get homologous by species id}
\usage{
get_homology_by_species_id(species, id)
}
\arguments{
\item{species}{A string representing the species name (e.g., "homo_sapiens").}

\item{id}{A string representing the Ensembl gene ID.}
}
\value{
A list of parsed JSON responses containing homology information
for the provided species and Ensembl gene ID.

See more about the implemented endpoint \code{\link[=get_homology_by_species_id]{get_homology_by_species_id()}}
on the following \href{https://rest.ensembl.org/documentation/info/homology_species_gene_id}{GET homology/id/:species/:id}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves homology information (orthologs) by species and Ensembl gene ID
}
\examples{
get_homology_by_species_id("homo_sapiens", "ENSG00000157764")
}
