% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_homology_by_symbol}
\alias{get_homology_by_symbol}
\title{Get homologous by symbol}
\usage{
get_homology_by_symbol(species, symbol)
}
\arguments{
\item{species}{A string representing the species name (e.g., "homo_sapiens").}

\item{symbol}{A string representing the gene symbol (e.g., "BRCA2").}
}
\value{
A list of parsed JSON responses containing homology information for the provided species and gene symbol.

See more about the implemented endpoint \code{\link[=get_homology_by_symbol]{get_homology_by_symbol()}}
on the following \href{https://rest.ensembl.org/documentation/info/homology_symbol}{GET homology/symbol/:species/:symbol}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves homology information (orthologs) by species and symbol
}
\examples{
get_homology_by_symbol("homo_sapiens", "BRCA2")
}
