% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{get_individuals}
\alias{get_individuals}
\title{Get individuals for a population}
\usage{
get_individuals(
  species_name = "homo_sapiens",
  population = "1000GENOMES:phase_3:CEU",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Examples: \code{'homo_sapiens'}
(human), \code{'ovis_aries'} (Domestic sheep) or \code{'capra_hircus'}
(Goat).}

\item{population}{Population name. Find the available populations for a given
species with \code{\link[ensemblr]{get_populations}}.}

\item{verbose}{Whether to be verbose about the http requests and respective
responses' status.}

\item{warnings}{Whether to show warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}} of 5 variables:
\describe{
\item{species_name}{Ensembl species name: this is the name used internally
by Ensembl to uniquely identify a species by name. It is the scientific
name but formatted without capitalisation and spacing converted with an
underscore, e.g., \code{'homo_sapiens'}.}
\item{population}{Population.}
\item{description}{Description of the population.}
\item{individual}{Individual identifier.}
\item{gender}{Individual gender.}
}
}
\description{
This function retrieves individual-level information. The data is returned as
a \code{\link[tibble]{tibble}} where each row is an individual of a given
species and the columns are metadata about each individual. See below under
section Value for details about each column. Use the function
\code{get_populations()} to discover the available populations for a species.
}
\section{Ensembl REST API endpoints}{
\code{get_individuals()} makes GET requests to
\href{https://rest.ensembl.org/documentation/info/variation_population_name}{/info/variation/populations/:species:/:population_name}.
}

\examples{
# Get human individuals for populaton "1000GENOMES:phase_3:CEU" (default)
get_individuals()

# Get Finnish individuals ("1000GENOMES:phase_3:FIN")
get_individuals(population = '1000GENOMES:phase_3:FIN')

}
