% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_karyotypes.R
\name{get_karyotypes}
\alias{get_karyotypes}
\title{Get the karyotype of a species}
\usage{
get_karyotypes(
  species_name = "homo_sapiens",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Examples: \code{'homo_sapiens'}
(human), \code{'ovis_aries'} (Domestic sheep) or \code{'capra_hircus'}
(Goat).}

\item{verbose}{Whether to be chatty.}

\item{warnings}{Whether to print warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}}, each row being a chromosome,
of 4 variables:
\describe{
\item{\code{species_name}}{Ensembl species name: this is the name used internally
by Ensembl to uniquely identify a species by name. It is the scientific
name but formatted without capitalisation and spacing converted with an
underscore, e.g., \code{'homo_sapiens'}.}
\item{\code{coord_system}}{Coordinate system type.}
\item{\code{chromosome}}{Chromosome name.}
\item{\code{length}}{Genomic length of the chromsome in base pairs.}
}
}
\description{
This function retrieves the set of chromosomes of a species.
}
\examples{
# Get the karyotype of Caenorhabditis elegans
get_karyotypes('caenorhabditis_elegans')

# Get the karyotype of the Giant panda
get_karyotypes('ailuropoda_melanoleuca')

}
