% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_ld_by_variant}
\alias{get_ld_by_variant}
\title{Get Linkage Disequilibrium (LD) values}
\usage{
get_ld_by_variant(
  species,
  id,
  population_name,
  attribs = NULL,
  callback = "randomlygeneratedname",
  d_prime = NULL,
  r2 = NULL,
  window_size = 500
)
}
\arguments{
\item{species}{A string representing the species name
or alias.}

\item{id}{A string representing the variant ID
(e.g., rs56116432).}

\item{population_name}{A string representing the population
for which to compute LD.
Use GET /info/variation/populations/:species?filter=LD to
retrieve a list of all populations with LD data. This endpoint is already
implemented. Please, use \code{\link[=get_species_populations]{get_species_populations()}} with filter "LD"
to retrieve valid populations.}

\item{attribs}{\emph{(Optional)} A boolean indicating whether
to add variation attributes such as chromosome, start, end, strand,
consequence type, and clinical significance. Default is NULL.}

\item{callback}{\emph{(Optional)} A string representing the name of the
callback subroutine for JSONP responses.}

\item{d_prime}{\emph{(Optional)} A float value (0-1) to filter results
by D' (linkage disequilibrium measure). Only returns pairs with D' ≥
the specified value. Default is NULL.}

\item{r2}{\emph{(Optional)} A float value (0-1) to filter results by
r² (correlation coefficient). Only returns pairs with r² ≥ the specified
value.. Default is NULL.}

\item{window_size}{\emph{(Optional)} An integer specifying the window size
in kb (max 500). Defaults to 500 kb.}
}
\value{
A list of parsed JSON responses containing the LD values for
the provided variant.
}
\description{
Computes and retrieves LD values between a given variant and all other
variants within a window of up to 500 kb in the specified population.
}
\note{
See more about the implemented endpoint \code{\link[=get_ld_by_variant]{get_ld_by_variant()}}
on the following \href{https://rest.ensembl.org/documentation/info/ld}{GET ld/:species/:id/:population_name}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\examples{
get_ld_by_variant("homo_sapiens", "rs56116432", "1000GENOMES:phase_3:KHV")
get_ld_by_variant("homo_sapiens", "rs56116432", "1000GENOMES:phase_3:KHV",
                  d_prime = 0.8, r2 = 0.85)
get_ld_by_variant("homo_sapiens", "rs56116432", "1000GENOMES:phase_3:KHV",
                  window_size = 250)

}
