% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_ld_values_by_region}
\alias{get_ld_values_by_region}
\title{Get Linkage Disequilibrium (LD) Values for a Genomic Region}
\usage{
get_ld_values_by_region(
  species,
  region,
  population_name,
  callback = "randomlygeneratedname",
  d_prime = NULL,
  r2 = NULL
)
}
\arguments{
\item{species}{A string representing the species name or alias
(e.g., "homo_sapiens").}

\item{region}{A string defining the genomic region in the format
"chr:start..end".
The maximum region size allowed is 500 kb. If the region overlaps
the MHC region, the maximum is 10 kb.}

\item{population_name}{A string representing the population for which
LD should be computed.
Use \code{\link[=get_species_populations]{get_species_populations()}} with filter "LD" to retrieve valid
populations.}

\item{callback}{\emph{(Optional)} A string representing the name of the
callback subroutine for JSONP responses.}

\item{d_prime}{\emph{(Optional)} A float value (0-1) to filter results
by D' (linkage disequilibrium measure).
Only returns pairs with D' ≥ the specified value.}

\item{r2}{\emph{(Optional)} A float value (0-1) to filter results by r²
(correlation coefficient).
Only returns pairs with r² ≥ the specified value.}
}
\value{
A parsed JSON response containing the LD values for all variant
pairs within the specified region.

See more about the implemented endpoint \code{\link[=get_ld_values_by_region]{get_ld_values_by_region()}}
on the following \href{https://rest.ensembl.org/documentation/info/ld_region}{GET ld/:species/region/:region/:population_name}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Computes and retrieves LD values between all pairs of variants within
a defined genomic region.
}
\examples{
get_ld_values_by_region(species = "homo_sapiens",
                       region = "6:25837556..25843455",
                       population_name = "1000GENOMES:phase_3:KHV")
get_ld_values_by_region(species = "homo_sapiens",
                       region = "6:25837556..25843455",
                       population_name = "1000GENOMES:phase_3:KHV",
                       r2 = 0.85)
get_ld_values_by_region(species = "homo_sapiens",
                       region = "6:25837556..25843455",
                       population_name = "1000GENOMES:phase_3:KHV",
                       d_prime = 1.0)
}
