% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_disequilibrium.R
\name{get_ld_variants_by_by_window}
\alias{get_ld_variants_by_by_window}
\alias{get_ld_variants_by_window}
\alias{get_ld_variants_by_pair}
\alias{get_ld_variants_by_range}
\title{Get linkage disequilibrium data for variants}
\usage{
get_ld_variants_by_window(variant_id, genomic_window_size = 500L,
  species_name = "homo_sapiens",
  population = "1000GENOMES:phase_3:CEU", d_prime = 0, r_squared = 0,
  verbose = FALSE, warnings = TRUE, progress_bar = TRUE)

get_ld_variants_by_pair(variant_id1, variant_id2,
  species_name = "homo_sapiens",
  population = "1000GENOMES:phase_3:CEU", d_prime = 0, r_squared = 0,
  verbose = FALSE, warnings = TRUE, progress_bar = TRUE)

get_ld_variants_by_range(genomic_range, species_name = "homo_sapiens",
  population = "1000GENOMES:phase_3:CEU", d_prime = 0, r_squared = 0,
  verbose = FALSE, warnings = TRUE, progress_bar = TRUE)
}
\arguments{
\item{variant_id}{A variant identifier, e.g., \code{'rs123'}. This argument
is to be used with function \code{get_ld_variants_by_window()}. Note that this
argument is not the same as \code{variant_id1} or \code{variant_id2}, to be
used with function \code{get_ld_variants_by_pair}.}

\item{genomic_window_size}{An integer vector specifying the genomic window
size in kilobases (kb) around the variant indicated in \code{variant_id}.
This argument is to be used with function
\code{get_ld_variants_by_window()}. At the moment, the Ensembl REST API
does not allow values greater than 500kb. A window size of 500 means
looking 250kb upstream and downstream the variant passed as
\code{variant_id}. The minimum value for this argument is \code{1L}, not
\code{0L}.}

\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Examples: \code{'homo_sapiens'}
(human), \code{'ovis_aries'} (Domestic sheep) or \code{'capra_hircus'} (Goat).}

\item{population}{Population for which to compute linkage disequilibrium. TODO.}

\item{d_prime}{\eqn{D'} is a measure of linkage disequilibrium.
\code{d_prime} defines a cut-off threshold: only variants whose
\eqn{D' \ge }\code{d_prime} are returned.}

\item{r_squared}{\eqn{r^2} is a measure of linkage disequilibrium.
\code{r_squared} defines a cut-off threshold: only variants whose
\eqn{r^2 \ge }\code{r_squared} are returned.}

\item{verbose}{Whether to be verbose about the http requests and respective
responses' status.}

\item{warnings}{Whether to show warnings.}

\item{progress_bar}{Whether to show a progress bar.}

\item{variant_id1}{The first variant of a pair of variants. Used with
\code{variant_id2}. Note that this argument is not the same as
\code{variant_id}. This argument is to be used with function
\code{get_ld_variants_by_pair()}.}

\item{variant_id2}{The second variant of a pair of variants. Used with
\code{variant_id1}. Note that this argument is not the same as
\code{variant_id}. This argument is to be used with function
\code{get_ld_variants_by_pair()}.}

\item{genomic_range}{Genomic range formatted as a string
\code{"chr:start..end"}, e.g., \code{"X:1..10000"}. Check function
\code{\link[ensemblr]{genomic_range}} to easily create these ranges from
vectors of start and end positions. This argument is to be used with
function \code{get_ld_variants_by_range()}.}
}
\value{
A \code{\link[tibble]{tibble}} of 6 variables:
\describe{
  \item{\code{species_name}}{Ensembl species name: this is the name used internally
  by Ensembl to uniquely identify a species by name. It is the scientific
  name but formatted without capitalisation and spacing converted with an
  underscore, e.g., \code{'homo_sapiens'}.}
  \item{\code{population}}{Population for which to compute linkage disequilibrium.}
  \item{\code{variant_id1}}{First variant identifier.}
  \item{\code{variant_id2}}{Second variant identifier.}
  \item{\code{d_prime}}{\eqn{D'} between the two variants.}
  \item{\code{r_squared}}{\eqn{r^2} between the two variants.}
}
}
\description{
Gets linkage disequilibrium data for variants from Ensembl REST API. There
are three ways to query, either by:
\describe{
\item{A genomic window centred on a
variant:}{\code{get_ld_variants_by_window(variant_id, genomic_window_size,
...)}}
\item{A pair of variants:}{\code{get_ld_variants_by_pair(variant_id1,
variant_id2, ...)}}
\item{A genomic range:}{\code{get_ld_variants_by_range(genomic_range, ...)}}
}
}
\examples{
# Retrieve variants in LD by a window size of 1kb:
# 1kb: 500 bp upstream and 500 bp downstream of variant.
get_ld_variants_by_window('rs123', genomic_window_size = 1L)

# Retrieve LD measures for pairs of variants:
get_ld_variants_by_pair(
  variant_id1 = c('rs123', 'rs35439278'),
  variant_id2 = c('rs122', 'rs35174522')
)

# Retrieve variants in LD within a genomic range
get_ld_variants_by_range('7:100000..100500')
}
