% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_pairwise_ld_values}
\alias{get_pairwise_ld_values}
\title{Get Pairwise Linkage Disequilibrium (LD) Values}
\usage{
get_pairwise_ld_values(
  species,
  id1,
  id2,
  callback = "randomlygeneratedname",
  d_prime = NULL,
  r2 = NULL,
  population_name = NULL
)
}
\arguments{
\item{species}{A string representing the species name or
alias (e.g., "homo_sapiens").}

\item{id1}{A string representing the first variant ID
(e.g., "rs6792369").}

\item{id2}{A string representing the second variant ID
(e.g., "rs1042779").}

\item{callback}{\emph{(Optional)} A string representing the name of
the callback
subroutine for JSONP responses.}

\item{d_prime}{\emph{(Optional)} A float value (0-1) to filter results
by D' (linkage disequilibrium measure).
Only returns pairs with D' ≥ the specified value.}

\item{r2}{\emph{(Optional)} A float value (0-1) to filter results by r²
(correlation coefficient). Only returns pairs with r² ≥ the specified value.}

\item{population_name}{\emph{(Optional)} A string representing the population
for which to compute LD. Use \code{\link[=get_species_populations]{get_species_populations()}} with filter "LD"
to retrieve valid populations.}
}
\value{
A parsed JSON response containing the LD values for the specified
variant pair.

See more about the implemented endpoint \code{\link[=get_pairwise_ld_values]{get_pairwise_ld_values()}}
on the following \href{https://rest.ensembl.org/documentation/info/ld_pairwise}{GET ld/:species/pairwise/:id1/:id2}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Computes and retrieves LD values between two given variants.
}
\examples{
get_pairwise_ld_values(species = "homo_sapiens", id1 = "rs6792369",
                      id2 = "rs1042779")
get_pairwise_ld_values(species = "homo_sapiens", id1 = "rs6792369",
                      id2 = "rs1042779", r2 = 0.85)
get_pairwise_ld_values(species = "homo_sapiens", id1 = "rs6792369",
                      id2 = "rs1042779", d_prime = 1.0,
                      population_name = "1000GENOMES:phase_3:KHV")
}
