% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_population_individuals}
\alias{get_population_individuals}
\title{Get Population Individuals}
\usage{
get_population_individuals(
  species,
  population_name,
  callback = "randomlygeneratedname"
)
}
\arguments{
\item{species}{(Required) A string representing the species name or alias (e.g., "human").}

\item{population_name}{(Required) A string representing the name of the population (e.g., "1000GENOMES:phase_3:ASW").}

\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}
}
\value{
A parsed JSON response containing the list of individuals for the specified population.

See more about the implemented endpoint \code{\link[=get_population_individuals]{get_population_individuals()}}
on the following \href{https://rest.ensembl.org/documentation/info/variation_population_name}{GET info/variation/populations/:species/:population_name}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of all individuals for a specified population from a species in Ensembl.
}
\examples{
get_population_individuals(species = "human", population_name = "1000GENOMES:phase_3:ASW")
get_population_individuals(species = "homo_sapiens", population_name = "1000GENOMES:phase_3:YRI")
get_population_individuals(
  species = "human",
  population_name = "1000GENOMES:phase_3:CEU",
  callback = "randomCallback"
)
}
