% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{get_populations}
\alias{get_populations}
\title{Get populations for a species}
\usage{
get_populations(species_name = "homo_sapiens", ld_only = TRUE,
  verbose = FALSE, warnings = TRUE, progress_bar = TRUE)
}
\arguments{
\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Examples: \code{'homo_sapiens'}
(human), \code{'ovis_aries'} (Domestic sheep) or \code{'capra_hircus'}
(Goat).}

\item{ld_only}{Whether to restrict populations returned to only populations
with linkage disequilibrium data.}

\item{verbose}{Whether to be verbose about the http requests and respective
responses' status.}

\item{warnings}{Whether to show warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}} of 4 variables:
\describe{
  \item{species_name}{Ensembl species name: this is the name used internally
  by Ensembl to uniquely identify a species by name. It is the scientific
  name but formatted without capitalisation and spacing converted with an
  underscore, e.g., \code{'homo_sapiens'}.}
  \item{population}{Population.}
  \item{description}{Description of the population.}
  \item{cohort_size}{Cohort sample size.}
}
}
\description{
This function retrieves population-level information. The data is returned as
a \code{\link[tibble]{tibble}} where each row is a population of a given
species and the columns are metadata about each population. See below under
section Value for details about each column. The parameter \code{ld_only} to
restrict populations returned to only populations with linkage disequilibrium
information.
}
\examples{
# Get all human populations with linkage disequilibrium data
get_populations(species_name = 'homo_sapiens', ld_only = TRUE)

# Get all human populations
get_populations(species_name = 'homo_sapiens', ld_only = FALSE)

}
