% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_region_info}
\alias{get_region_info}
\title{Get information about a specific toplevel sequence region for a species}
\usage{
get_region_info(
  species,
  region_name,
  bands = 0,
  synonyms = 0,
  callback = "randomlygeneratedname"
)
}
\arguments{
\item{species}{A string representing the species name or alias (e.g., "homo_sapiens").}

\item{region_name}{A string representing the name of the toplevel sequence region (e.g., "X").}

\item{bands}{(Optional) A boolean (0 or 1) indicating whether to include
karyotype band information. Default is 0.}

\item{synonyms}{(Optional) A boolean (0 or 1) indicating whether to include
information about known synonyms. Default is 0.}

\item{callback}{(Optional) A string representing the name of the callback
subroutine for JSONP responses.}
}
\value{
A list of parsed JSON responses containing information about
the specified sequence region for the given species.

See more about the implemented endpoint \code{\link[=get_region_info]{get_region_info()}}
on the following \href{https://rest.ensembl.org/documentation/info/assembly_stats}{GET info/assembly/:species/:region_name}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves information about the specified toplevel sequence region for
a given species, with optional details on karyotype bands and synonyms.
}
\examples{
get_region_info("homo_sapiens", "X")
get_region_info("homo_sapiens", "X", bands = 1)
get_region_info("homo_sapiens", "X", synonyms = 1, callback = "randomlygeneratedname")
}
