% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versioning.R
\name{get_rest_version}
\alias{get_rest_version}
\title{Retrieve the current version of the Ensembl REST API}
\usage{
get_rest_version(verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{verbose}{Whether to be chatty.}

\item{warnings}{Whether to print warnings.}
}
\value{
A scalar character vector with Ensembl REST API version.
}
\description{
Retrieve the current version of the Ensembl REST API
}
