% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-versioning.R
\name{get_rest_version2}
\alias{get_rest_version2}
\title{Get REST API version}
\source{
\url{https://github.com/Ensembl/ensembl-rest/wiki/API-Versioning}
}
\usage{
get_rest_version2()
}
\value{
A numeric value representing the REST API version components:
\code{major}, \code{minor}, and \code{point}.
}
\description{
Retrieve the version of the Ensembl REST API currently in use.
The version format is \code{major.minor.point}.
}
\examples{
\dontrun{
rest_version <- get_rest_version()
print(rest_version)
}
}
