% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versioning.R
\name{get_software_version}
\alias{get_software_version}
\title{Retrieve the Perl API version}
\usage{
get_software_version(verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{verbose}{Whether to be chatty.}

\item{warnings}{Whether to print warnings.}
}
\value{
A scalar integer vector with the Perl API version.
}
\description{
Retrieve the Perl API version
}
