% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-versioning.R
\name{get_software_version2}
\alias{get_software_version2}
\title{Get API software version}
\source{
\url{https://github.com/Ensembl/ensembl-rest/wiki/API-Versioning}
}
\usage{
get_software_version2()
}
\value{
A numeric value representing the software version.
}
\description{
Retreive the version of the Ensembl software the REST API is using.
}
\examples{
\dontrun{
software_version <- get_software_version()
print(software_version)
}
}
