% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{get_species}
\alias{get_species}
\title{Get Ensembl species}
\usage{
get_species(
  division = get_divisions(),
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{division}{Ensembl division, e.g., \code{"EnsemblVertebrates"} or
\code{"EnsemblBacteria"}, or a combination of several divisions. Check
function \code{\link[ensemblr]{get_divisions}} to get available Ensembl
divisions.}

\item{verbose}{Whether to be verbose about the http requests and respective
responses' status.}

\item{warnings}{Whether to show warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}} of 12 variables:
\describe{
  \item{division}{Ensembl division: \code{"EnsemblVertebrates"},
  \code{"EnsemblMetazoa"}, \code{"EnsemblPlants"}, \code{"EnsemblProtists"},
  \code{"EnsemblFungi"} or \code{"EnsemblBacteria"}.}
  \item{taxon_id}{NCBI taxon identifier.}
  \item{species_name}{Ensembl species name: this is the name used internally
  by Ensembl to uniquely identify a species by name. It is the scientific
  name but formatted without capitalisation and spacing converted with an
  underscore, e.g., \code{'homo_sapiens'}.}
  \item{species_display_name}{Species display name: the name used for display
  on Ensembl website.}
  \item{species_common_name}{Species common name.}
  \item{release}{Ensembl release version.}
  \item{genome_assembly_name}{Code name of the genome assembly.}
  \item{genbank_assembly_accession}{Genbank assembly accession identifier.}
  \item{strain}{Species strain.}
  \item{strain_collection}{Species strain collection.}
  \item{species_aliases}{Other names or acronyms used to refer to the
  species. Note that this column is of the list type.}
  \item{groups}{Ensembl databases for which data exists for this species.
  Note that this column is of the list type.}
}
}
\description{
This function retrieves species-level information. The data is returned as a
\code{\link[tibble]{tibble}} where each row is a species and the columns are
metadata about each species. See below under section Value for details about
each column.
}
