% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_species_info}
\alias{get_species_info}
\title{Get Species Information}
\usage{
get_species_info(
  callback = "randomlygeneratedname",
  division = "EnsemblVertebrates",
  hide_strain_info = 0,
  strain_collection = NULL
)
}
\arguments{
\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}

\item{division}{(Optional) A string to filter by Ensembl or Ensembl Genomes division (default is "EnsemblVertebrates").}

\item{hide_strain_info}{(Optional) A boolean flag to show/hide strain and
strain_collection information (default is 0, which shows strain info).}

\item{strain_collection}{(Optional) A string to filter by strain collection (e.g., "mouse").}
}
\value{
A parsed JSON response containing species information.

See more about the implemented endpoint \code{\link[=get_species_info]{get_species_info()}}
on the following \href{https://rest.ensembl.org/documentation/info/species}{GET info/species}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of all available species, their aliases, available adaptor groups, and data release.
}
\examples{
get_species_info()
get_species_info(division = "EnsemblPlants")
get_species_info(hide_strain_info = 1)
get_species_info(strain_collection = "mouse")
}
