% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_species_populations}
\alias{get_species_populations}
\title{Get Populations for a Species}
\usage{
get_species_populations(
  species,
  callback = "randomlygeneratedname",
  filter = NULL
)
}
\arguments{
\item{species}{(Required) A string representing the species name or alias (e.g., "homo_sapiens").}

\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}

\item{filter}{(Optional) A string to restrict populations returned
(e.g., "LD" to filter populations with linkage disequilibrium data).}
}
\value{
A parsed JSON response containing the list of populations for the specified species.

See more about the implemented endpoint \code{\link[=get_species_populations]{get_species_populations()}}
on the following \href{https://rest.ensembl.org/documentation/info/variation_populations}{GET info/variation/populations/:species}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves a list of all populations for a specified species in Ensembl.
}
\examples{
get_species_populations(species = "homo_sapiens")
get_species_populations(species = "human", filter = "LD")
get_species_populations(
  species = "homo_sapiens",
  callback = "randomlygeneratedname",
  filter = "LD"
)
}
