% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_toplevel_sequence_info.R
\name{get_toplevel_sequence_info}
\alias{get_toplevel_sequence_info}
\title{Get toplevel sequences details}
\usage{
get_toplevel_sequence_info(
  species_name = "homo_sapiens",
  toplevel_sequence = c(1:22, "X", "Y", "MT"),
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Examples: \code{'homo_sapiens'}
(human), \code{'ovis_aries'} (Domestic sheep) or \code{'capra_hircus'}
(Goat).}

\item{toplevel_sequence}{A toplevel sequence name, e.g. chromosome names such
as \code{"1"}, \code{"X"}, or \code{"Y"}, or a non-chromosome sequence, e.g., a scaffold
such as \code{"KI270757.1"}.}

\item{verbose}{Whether to be chatty.}

\item{warnings}{Whether to print warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}}, each row being a toplevel sequence,
of 8 variables:
\describe{
\item{\code{species_name}}{Ensembl species name: this is the name used internally
by Ensembl to uniquely identify a species by name. It is the scientific name
but formatted without capitalisation and spacing converted with an
underscore, e.g., \code{'homo_sapiens'}.}
\item{\code{toplevel_sequence}}{Name of the toplevel sequence.}
\item{\code{is_chromosome}}{A logical indicating whether the toplevel sequence is
a chromosome (\code{TRUE}) or not (\code{FALSE}).}
\item{\code{coord_system}}{Coordinate system type.}
\item{\code{assembly_exception_type}}{Coordinate system type.}
\item{\code{is_circular}}{A logical indicating whether the toplevel sequence is a
circular sequence (\code{TRUE}) or not (\code{FALSE}).}
\item{\code{assembly_name}}{Assembly name.}
\item{\code{length}}{Genomic length toplevel sequence in base pairs.}
}
}
\description{
This function retrieves a few extra details about a toplevel sequence. These
sequences correspond to genomic regions in the genome assembly that are not a
component of another sequence region. Thus, toplevel sequences will be
chromosomes and any unlocalised or unplaced scaffolds.
}
\examples{
# Get details about human chromosomes (default)
get_toplevel_sequence_info()

# Get details about a scaffold
# (To find available toplevel sequences to query use the function
# `get_toplevel_sequences()`)
get_toplevel_sequence_info(species_name = 'homo_sapiens', toplevel_sequence = 'KI270757.1')

}
\seealso{
\code{\link[=get_toplevel_sequences]{get_toplevel_sequences()}}
}
