% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_toplevel_sequences.R
\name{get_toplevel_sequences}
\alias{get_toplevel_sequences}
\title{Get toplevel sequences by species}
\usage{
get_toplevel_sequences(
  species_name = "homo_sapiens",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Examples: \code{'homo_sapiens'}
(human), \code{'ovis_aries'} (Domestic sheep) or \code{'capra_hircus'}
(Goat).}

\item{verbose}{Whether to be chatty.}

\item{warnings}{Whether to print warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}}, each row being a toplevel sequence,
of 4 variables:
\describe{
\item{\code{species_name}}{Ensembl species name: this is the name used internally
by Ensembl to uniquely identify a species by name. It is the scientific
name but formatted without capitalisation and spacing converted with an
underscore, e.g., \code{'homo_sapiens'}.}
\item{\code{coord_system}}{Coordinate system type.}
\item{\code{toplevel_sequence}}{Name of the toplevel sequence.}
\item{\code{length}}{Genomic length toplevel sequence in base pairs.}
}
}
\description{
This function retrieves toplevel sequences. These sequences correspond to
genomic regions in the genome assembly that are not a component of another
sequence region. Thus, toplevel sequences will be chromosomes and any
unlocalised or unplaced scaffolds.
}
\examples{
# Get toplevel sequences for the human genome (default)
get_toplevel_sequences()

# Get toplevel sequences for Caenorhabditis elegans
get_toplevel_sequences('caenorhabditis_elegans')

}
