% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_sources.R
\name{get_variation_sources}
\alias{get_variation_sources}
\title{Retrieve variant sources}
\usage{
get_variation_sources(
  species_name = "human",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Examples: \code{'homo_sapiens'}
(human), \code{'ovis_aries'} (Domestic sheep) or \code{'capra_hircus'}
(Goat).}

\item{verbose}{Whether to be chatty.}

\item{warnings}{Whether to print warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}}, each row being a variant database,
of 8 variables:
\describe{
\item{species_name}{Ensembl species name: this is the name used internally
by Ensembl to uniquely identify a species by name. It is the scientific
name but formatted without capitalisation and spacing converted with an
underscore, e.g., \code{'homo_sapiens'}.}
\item{db_name}{Database name.}
\item{type}{Database type, e.g., \code{chip} (genotyping chip) or \code{lsdb}
(locus-specific database).}
\item{version}{Database version.}
\item{somatic_status}{Somatic status.}
\item{description}{Database description.}
\item{url}{Database's URL.}
\item{data_types}{Data types to be found at database.}
}
}
\description{
This function retrieves variant sources, i.e. a list of databases used by
Ensembl from which variant information is retrieved.
}
\section{Ensembl REST API endpoints}{
\code{get_variation_sources} makes GET requests to
\href{https://rest.ensembl.org/documentation/info/variation}{info/variation/:species}.
}

\examples{
# Retrieve variant sources for human (default)
get_variation_sources()

# Retrieve variant sources for mouse
get_variation_sources(species_name = 'mus_musculus')

}
