% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_references.R
\name{get_xrefs_by_ensembl_id}
\alias{get_xrefs_by_ensembl_id}
\title{Get cross-references by Ensembl ID}
\usage{
get_xrefs_by_ensembl_id(
  species_name,
  ensembl_id,
  all_levels = FALSE,
  ensembl_db = "core",
  external_db = "",
  feature = "",
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{species_name}{The species name, i.e., the scientific name, all letters
lowercase and space replaced by underscore. Examples: \code{'homo_sapiens'}
(human), \code{'ovis_aries'} (Domestic sheep) or \code{'capra_hircus'}
(Goat).}

\item{ensembl_id}{An Ensembl stable identifier, e.g. \code{"ENSG00000248234378"}.}

\item{all_levels}{A \code{logical} vector. Set to find all genetic features linked
to the stable ID, and fetch all external references for them. Specifying
this on a gene will also return values from its transcripts and
translations.}

\item{ensembl_db}{Restrict the search to an Ensembl database: typically one
of \code{'core'}, \code{'rnaseq'}, \code{'cdna'}, \code{'funcgen'} and \code{'otherfeatures'}.}

\item{external_db}{External database to be filtered by. By default no
filtering is applied.}

\item{feature}{Restrict search to a feature type: gene (\code{'gene'}), exon
(\code{'exon'}), transcript (\code{'transcript'}), and protein
(\code{'translation'}).}

\item{verbose}{Whether to be verbose about the http requests and respective
responses' status.}

\item{warnings}{Whether to show warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}} of 12
variables:
\describe{
\item{\code{species_name}}{Ensembl species name: this is the name used
internally by Ensembl to uniquely identify a species by name. It is the
scientific name but formatted without capitalisation and spacing converted
with an underscore, e.g., \code{'homo_sapiens'}.}
\item{\code{ensembl_id}}{An Ensembl stable identifier, e.g. \code{"ENSG00000248234378"}.}
\item{\code{ensembl_db}}{Ensembl database.}
\item{\code{primary_id}}{Primary identification in external database.}
\item{\code{display_id}}{Display identification in external database.}
\item{\code{external_db_name}}{External database name.}
\item{\code{external_db_display_name}}{External database display name.}
\item{\code{version}}{TODO}
\item{\code{info_type}}{There are two types of external cross references (XRef):
direct (\code{'DIRECT'}) or dependent (\code{'DEPENDENT'}). A direct cross
reference is one that can be directly linked to a gene, transcript or
translation object in Ensembl Genomes by synonymy or sequence similarity. A
dependent cross reference is one that is transitively linked to the object
via the direct cross reference. The value can also be \code{'UNMAPPED'} for
unmapped cross references, or \code{'PROJECTION'} for TODO.}
\item{\code{info_text}}{TODO}
\item{\code{synonyms}}{Other names or acronyms used to refer to the
the external database entry. Note that this column is of the list type.}
\item{\code{description}}{Brief description of the external database entry.}
}
}
\description{
This function retrieves cross-references to external databases by Ensembl
identifier. The data is returned as a \code{\link[tibble]{tibble}} where each
row is a cross reference related to the provided Ensembl identifier. See
below under section Value for details about each column.
}
