% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_xrefs_by_id}
\alias{get_xrefs_by_id}
\title{Get external linked references by id}
\usage{
get_xrefs_by_id(id)
}
\arguments{
\item{id}{A string representing the Ensembl Identifier (e.g., "ENSG00000157764").}
}
\value{
A list of parsed JSON responses containing external references
for the provided Ensembl identifier.

See more about the implemented endpoint \code{\link[=get_xrefs_by_id]{get_xrefs_by_id()}}
on the following \href{https://rest.ensembl.org/documentation/info/xref_id}{GET xrefs/id/:id}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Performs lookups of Ensembl Identifiers and retrieves their external
references in other databases
}
\examples{
get_xrefs_by_id("ENSG00000157764")
}
