% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_xrefs_by_name}
\alias{get_xrefs_by_name}
\title{Get external linked references by name}
\usage{
get_xrefs_by_name(species, name)
}
\arguments{
\item{species}{A string representing the species name (e.g., "homo_sapiens").}

\item{name}{A string representing the primary accession or display label
of the external reference.}
}
\value{
A list of parsed JSON responses containing information about
the provided external reference.

See more about the implemented endpoint \code{\link[=get_xrefs_by_name]{get_xrefs_by_name()}}
on the following \href{https://rest.ensembl.org/documentation/info/xref_name}{GET xrefs/name/:species/:name}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Performs a lookup based upon the primary accession or display label
of an external reference
}
\examples{
get_xrefs_by_name("homo_sapiens", "P38398")
}
