% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_xrefs_by_symbol}
\alias{get_xrefs_by_symbol}
\title{Get external linked references by symbol}
\usage{
get_xrefs_by_symbol(species, symbol)
}
\arguments{
\item{species}{A string representing the species name (e.g., "homo_sapiens").}

\item{symbol}{A string representing the external symbol (e.g., "BRCA2").}
}
\value{
A list of parsed JSON responses containing Ensembl objects linked
to the provided external symbol.

See more about the implemented endpoint \code{\link[=get_xrefs_by_symbol]{get_xrefs_by_symbol()}}
on the following \href{https://rest.ensembl.org/documentation/info/xref_external}{GET xrefs/symbol/:species/:symbol}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Looks up an external symbol and returns all Ensembl objects linked to it
}
\details{
This can be a display name for a gene/transcript/translation, a synonym,
or an externally linked reference.
If a gene's transcript is linked to the supplied symbol, the service will
return both gene and transcript (it supports transient links).
}
\examples{
get_xrefs_by_symbol("homo_sapiens", "BRCA2")
}
