% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_server.R
\name{is_ensembl_reachable}
\alias{is_ensembl_reachable}
\title{Is the Ensembl REST API server reachable?}
\usage{
is_ensembl_reachable(url = ensembl_server(), port = 443L,
  verbose = FALSE)
}
\arguments{
\item{url}{Ensembl REST API server URL. Default is
\url{https://rest.ensembl.org}. You should not need to change this parameter.}

\item{port}{Network port on which to ping the server. You should not need to
change this parameter.}

\item{verbose}{Whether to be verbose (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
A logical value: \code{TRUE} if EBI server is reachable, \code{FALSE}
  otherwise.
}
\description{
Check if the Ensembl server where REST API service is running is reachable.
This function attempts to connect to
\href{https://rest.ensembl.org}{https://rest.ensembl.org}, returning
\code{TRUE} on success, and \code{FALSE} otherwise. Set \code{verbose = TRUE}
for a step by step description of the connection attempt.
}
\examples{
# Check if the Ensembl Server is reachable
is_ensembl_reachable() # Returns TRUE or FALSE.

# Check if the GWAS Catalog Server is reachable
# and show exactly at what step is it failing (if that is the case)
is_ensembl_reachable(verbose = TRUE)

}
