% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pairwise_combn}
\alias{pairwise_combn}
\title{Pairwise combinations}
\usage{
pairwise_combn(x)
}
\arguments{
\item{x}{A character vector from which to generate the combinations.}
}
\value{
A \code{\link[tibble]{tibble}} of two columns where each row is a
pairwise combination.
}
\description{
Generates pairwise combinations from the supplied vector. Never returns the
pairs self against self.
}
\keyword{internal}
