% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{ping_service}
\alias{ping_service}
\title{Check Service Status}
\usage{
ping_service(callback = "randomlygeneratedname")
}
\arguments{
\item{callback}{(Optional) A string representing the name of the callback
subroutine for JSONP responses.}
}
\value{
A parsed JSON response indicating the status of the service.

See more about the implemented endpoint \code{\link[=ping_service]{ping_service()}}
on the following \href{https://rest.ensembl.org/documentation/info/ping}{GET info/ping}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Sends a ping request to the server to check if the service is alive.
}
\examples{
ping_service()
ping_service(callback = "randomlygeneratedname")
}
