% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Request an endpoint from Ensembl REST API}
\usage{
request(resource_url, base_url = ensembl_server(), verbose = FALSE,
  warnings = TRUE)
}
\arguments{
\item{resource_url}{Endpoint URL. The endpoint is internally appended to the
\code{base_url}. It should start with a forward slash (\code{'/'}).}

\item{base_url}{The Ensembl REST API base URL.}

\item{verbose}{Whether to be verbose.}

\item{warnings}{Whether to print warnings.}
}
\value{
A named list of four elements:
\describe{
\item{url}{The URL endpoint.}
\item{response_code}{\href{https://tinyurl.com/8yqvhwf}{HTTP
status code}.}
\item{status}{A string describing the status of the response obtained:
\code{"OK"} if successful or a description the error.}
\item{content}{The parsed JSON as a nested list, as returned by
\code{\link[jsonlite]{fromJSON}}.}
}
}
\description{
Performs a \code{\link[httr]{GET}} request on the endpoint as specified by
\code{resource_url}.
}
