% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request_parallel}
\alias{request_parallel}
\title{Parallel version of request}
\usage{
request_parallel(
  resource_urls,
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{resource_urls}{Vector of endpoint URLs. Each endpoint is internally
appended to the \code{base_url}. It should start with a forward slash
(\code{'/'}).}

\item{verbose}{Whether to be verbose.}

\item{warnings}{Whether to print warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A list of named lists of four elements:
\describe{
\item{url}{The URL endpoint.}
\item{response_code}{\href{https://tinyurl.com/8yqvhwf}{HTTP
status code}.}
\item{status}{A string describing the status of the response obtained:
\code{"OK"} if successful or a description the error.}
\item{content}{The parsed JSON as a nested list, as returned by
\code{\link[jsonlite]{fromJSON}}.}
}
}
\description{
Performs a \code{\link[httr]{GET}} request on each of the endpoints as
specified by \code{resource_urls}.
}
\keyword{internal}
