% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rest_api_endpoints}
\alias{rest_api_endpoints}
\title{Ensembl REST API Endpoints.}
\format{A data frame with 118 rows and 4 variables:
\describe{
  \item{section}{Section.}
  \item{endpoint}{Ensembl REST API endpoint.}
  \item{description}{A short description of the resource.}
  \item{last_update_date}{Time stamp of last time this dataset was
  downloaded from Ensembl.}
}}
\source{
\url{https://rest.ensembl.org/}
}
\usage{
rest_api_endpoints
}
\description{
A dataset containing the Ensembl REST API endpoints, as listed in
\url{https://rest.ensembl.org/}.
}
\keyword{datasets}
