% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{strand_to_code}
\alias{strand_to_code}
\title{Convert strand values to integer codes}
\usage{
strand_to_code(strand, .default = NA_integer_, .missing = NA_integer_)
}
\arguments{
\item{strand}{Character vector of strandness values: \code{'forward'} or
\code{'reverse'}.}

\item{.default}{What value to use by default for values other than
\code{'forward'} or \code{'reverse'}.}

\item{.missing}{What value to use when recoding \code{NA_character_}.}
}
\value{
An integer vector.
}
\description{
This function converts a character vector of values \code{'forward'} or
\code{'reverse'} to \code{1} or \code{-1}, respectively.
}
\seealso{
\code{\link{code_to_strand}}
}
\keyword{internal}
