% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{warn_when_request_errored}
\alias{warn_when_request_errored}
\title{Warn if response errored}
\usage{
warn_when_request_errored(response)
}
\arguments{
\item{response}{A \code{\link[httr]{response}} object.}
}
\value{
A scalar character vector with a warning message, or the string
\code{'OK'} if the response was successful, although this function is
called mostly for its side effect, i.e., the triggering of a warning.
}
\description{
Warn if an httr \code{\link[httr]{response}} errored. It also returns a tidy
warning message.
}
\keyword{internal}
