% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get}
\alias{get}
\title{The function for the GET method}
\usage{
get(res, ..., .headers = req_headers(), rate = 15/60)
}
\arguments{
\item{res}{The resource (path) for the API request, can include variables
in curly braces \code{{}} that will be replaced with the corresponding parameter.}

\item{...}{Additional named parameters to be included in the request URL.}

\item{.headers}{An S3 list with class \code{ensemblr_req_hdr}. Use the helper
\code{\link[=req_headers]{req_headers()}} to create such an object.}

\item{rate}{The maximum number of requests per second to allow.
Defaults to 15 per minute (15/60).}
}
\value{
A list of responses, one for each request made.
}
\description{
The \code{\link[=get]{get()}} function is a wrapper around the \code{\link[=reqs]{reqs()}} function that performs
GET requests to the Ensembl API, handling rate limiting automatically.
}
\keyword{internal}
