% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_cafe_genetree_by_symbol}
\alias{get_cafe_genetree_by_symbol}
\title{Get cafe gene tree by symbol}
\usage{
get_cafe_genetree_by_symbol(species, symbol)
}
\arguments{
\item{species}{A string representing the species name (e.g.,
"homo_sapiens").}

\item{symbol}{A string representing the gene symbol (e.g., "BRCA2").}
}
\value{
A list of parsed JSON responses containing the cafe tree
for the provided species and gene symbol.
}
\description{
Retrieves the cafe tree of the gene tree that contains the gene identified
by a symbol
}
\note{
See more about the implemented endpoint \code{\link[=get_cafe_genetree_by_symbol]{get_cafe_genetree_by_symbol()}}
on the following \href{https://rest.ensembl.org/documentation/info/cafe_tree_member_symbol}{GET cafe/genetree/member/symbol/:species/:symbol}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\examples{
get_cafe_genetree_by_symbol("homo_sapiens", "BRCA2")
}
