% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-versioning.R
\name{get_data_version2}
\alias{get_data_version2}
\title{Get data version}
\source{
\url{https://github.com/Ensembl/ensembl-rest/wiki/API-Versioning}
}
\usage{
get_data_version2()
}
\value{
A numeric vector of data release versions.
}
\description{
Retreive the version(s) of the Ensembl data that the REST API is accessing.
}
\examples{
\dontrun{
data_version <- get_data_version()
print(data_version)
}
}
