% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_genome_info_by_assembly}
\alias{get_genome_info_by_assembly}
\title{Get genome information by assembly ID}
\usage{
get_genome_info_by_assembly(
  assembly_id,
  callback = "randomlygeneratedname",
  expand = NULL
)
}
\arguments{
\item{assembly_id}{(Required) A string representing the INSDC assembly ID (optionally versioned, e.g., "GCA_902167145.1").}

\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}

\item{expand}{(Optional) A boolean value (0 or 1). If set to 1, expands the information
to include details of sequences (can be very large).}
}
\value{
A parsed JSON response containing information about the genome associated
with the specified assembly ID.

See more about the implemented endpoint \code{\link[=get_genome_info_by_assembly]{get_genome_info_by_assembly()}}
on the following \href{https://rest.ensembl.org/documentation/info/info_genomes_assembly}{GET info/genomes/assembly/:assembly_id}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves information about a genome associated with a specified assembly ID.
}
\examples{
get_genome_info_by_assembly(assembly_id = "GCA_902167145.1")
get_genome_info_by_assembly(assembly_id = "GCA_902167145.1", expand = 1)
get_genome_info_by_assembly(assembly_id = "GCA_902167145.1", callback = "randomlygeneratedname")
}
