% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-endpoints.R
\name{get_genome_info_by_division}
\alias{get_genome_info_by_division}
\title{Get genome information for a specific division}
\usage{
get_genome_info_by_division(
  division_name,
  callback = "randomlygeneratedname",
  expand = NULL
)
}
\arguments{
\item{division_name}{(Required) A string representing the name of the division (e.g., "EnsemblPlants").}

\item{callback}{(Optional) A string representing the name of the callback subroutine for JSONP responses.}

\item{expand}{(Optional) A boolean value (0 or 1). If set to 1, expands the information to include details of sequences (can be very large).}
}
\value{
A parsed JSON response containing information about genomes in the specified division.

See more about the implemented endpoint \code{\link[=get_genome_info_by_division]{get_genome_info_by_division()}}
on the following \href{https://rest.ensembl.org/documentation/info/info_genomes_division}{GET info/genomes/division/:division_name}
from the official \href{https://rest.ensembl.org/}{Ensembl Rest API}.
}
\description{
Retrieves information about all genomes in a given division.
Note: The response may be very large for divisions like Ensembl Bacteria.
}
\examples{
get_genome_info_by_division(division_name = "EnsemblPlants")
get_genome_info_by_division(division_name = "EnsemblPlants", expand = 1)
get_genome_info_by_division(division_name = "EnsemblPlants", callback = "randomlygeneratedname")
}
