% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{get_id}
\alias{get_id}
\title{Get details about an Ensembl identifier}
\usage{
get_id(id, verbose = FALSE, warnings = TRUE, progress_bar = TRUE)
}
\arguments{
\item{id}{A character vector of Ensembl identifiers. Ensembl identifiers have
the form ENS[species prefix][feature type prefix][a unique eleven digit
number]. \code{id} should not contain NAs. Please note that while
\code{'ENSG00000157764'} is a valid identifier as a query,
\code{'ENSG00000157764.13'} is not.}

\item{verbose}{Whether to be verbose about the http requests and respective
responses' status.}

\item{warnings}{Whether to show warnings.}

\item{progress_bar}{Whether to show a progress bar.}
}
\value{
A \code{\link[tibble]{tibble}} of 9 variables:
\describe{
\item{\code{id}}{Ensembl identifier.}
\item{\code{id_latest}}{Ensembl identifier including the version suffix.}
\item{\code{type}}{Entity type: gene (\code{'Gene'}), exon (\code{'Exon'}),
transcript (\code{'Transcript'}), and protein (\code{'Translation'}).}
\item{\code{id_version}}{Ensembl identifier version, indicates how many
times that entity has changed during its time in Ensembl.}
\item{\code{release}}{Ensembl release version.}
\item{\code{is_current}}{Is this the latest identifier for the represented entity.}
\item{\code{genome_assembly_name}}{Code name of the genome assembly.}
\item{\code{peptide}}{TODO}
\item{\code{possible_replacement}}{TODO}
}
}
\description{
This function retrieves information about one or more Ensembl identifiers.
Ensembl identifiers for which information is available are: genes, exons,
transcripts and proteins.
}
\examples{
get_id(c('ENSDARE00000830915', 'ENSG00000248378', 'ENSDART00000033574', 'ENSP00000000233'))

}
