% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post}
\alias{post}
\title{The function for POST method}
\usage{
post(res, ..., .headers = req_headers(), .body, rate = 15/60)
}
\arguments{
\item{res}{The resource (path) for the API request, can include variables
in curly braces \code{{}} that will be replaced with the corresponding
parameter.}

\item{...}{Additional named parameters to be included in the request URL.}

\item{.headers}{An S3 list with class \code{ensemblr_req_hdr}. Use the helper
\code{\link[=req_headers]{req_headers()}} to create such an object.}

\item{.body}{The body of the POST request, can be a string or raw vector.}

\item{rate}{The maximum number of requests per second to allow.
on a request (default: 5).
Defaults to 15 per minute (15/60).}
}
\value{
A list of responses, one for each request made.
}
\description{
The \code{\link[=post]{post()}} function is a wrapper around the \code{\link[=reqs]{reqs()}} function that
performs POST requests to the Ensembl API, handling rate limiting
automatically.
}
\keyword{internal}
